+++ app/Http/Controllers/LaboratorioController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Laboratorio;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class LaboratorioController extends Controller
 {
@@ -19,6 +20,7 @@ class LaboratorioController extends Controller
     }
 
     public function storeAdd(Request $request){
+        abort_if(Gate::denies('laboratory_create'), 403);
         Laboratorio::create(['nombreLaboratorio' => $request->nombreLaboratorio]);
         toast('<p class="font-weight-light text-dark">Laboratorio agregado correctamente.</p>','success')
         ->toHtml()
@@ -27,6 +29,7 @@ class LaboratorioController extends Controller
     }
 
     public function update(Request $request, $id){
+        abort_if(Gate::denies('category_edit'), 403);
         $laboratory = Laboratorio::findOrfail($id);
         $laboratory->update(['nombreLaboratorio' => $request->nombreLaboratorio]);
         toast('<p class="font-weight-light text-dark">Laboratorio actualizado correctamente.</p>','success')
@@ -36,6 +39,7 @@ class LaboratorioController extends Controller
     }
 
     public function delete($id){
+        abort_if(Gate::denies('category_U_delete'), 403);
         $laboratory = Laboratorio::findOrfail($id);
         $laboratory->delete();
         toast('<p class="font-weight-light text-dark">Laboratorio eliminado</p>','success')
